% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{mn_pal}
\alias{mn_pal}
\title{MN palettes with ramped colours}
\usage{
mn_pal(palette = "mn_qual", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'mn_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
MN palettes with ramped colours
}
\examples{
library(scales)
show_col(mn_pal()(10))

filled.contour(volcano,color.palette = mn_pal(), asp=1)

}
