% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{mn_palettes}
\alias{mn_palettes}
\title{Minnesota Color Palettes}
\format{
An object of class \code{list} of length 13.
}
\usage{
mn_palettes
}
\description{
A collection of color palettes based on the official state brand colors and others inspired by the landscape and wildlife of Minnesota.
}
\examples{

# Make an x-y plot using the Primary palette
library(tidyverse)

df <- data.frame(x = rnorm(100, 0, 20),
                y = rnorm(100, 0, 20),
               cl = sample(letters[1:8], 100, replace=T))
ggplot(df, aes(x, y, color = cl, shape=cl)) +
  geom_point(size=4) + 
  scale_color_mncolors(palette = "primary") +
  theme_bw()

# Make a histogram using the Secondary palette
ggplot(df, aes(x, fill=cl)) + 
  geom_histogram() +
  scale_fill_mncolors(palette = "secondary")

}
\keyword{datasets}
