% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_mn}
\alias{scale_color_mn}
\alias{scale_colour_mn}
\title{Setup colour palette for ggplot2}
\usage{
scale_color_mn(
  palette = "primary_extended",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)

scale_colour_mn(
  palette = "primary_extended",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'mn_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colours?}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
Setup colour palette for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl))) +
  scale_colour_mn(palette="mn_qual")
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = hp)) +
  scale_colour_mn(palette="mn_qual", discrete = FALSE)
ggplot(data = mpg) +
  geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
  scale_colour_mn(palette="accent")
ggplot(diamonds) + geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_mn()
}
