% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_mn}
\alias{scale_fill_mn}
\title{#' Setup fill palette for ggplot2}
\usage{
scale_fill_mn(
  ...,
  palette = "mn_qual",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'mn_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{pass through parameter to \code{viridis}}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
#' Setup fill palette for ggplot2
}
