% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{viz_palette}
\alias{viz_palette}
\title{Show a palette}
\usage{
viz_palette(pal, ttl = deparse(substitute(pal)), num = length(pal))
}
\arguments{
\item{pal}{character, vector of (hexadecimal) colors representing a palette}

\item{ttl}{character, title to be displayed (the name of the palette)}

\item{num}{numeric, the number of colors to display}
}
\description{
Display a single palette to see whether it meets your needs.
If no \code{num} parameter is given,
all the colors in the palette will be displayed.
If \code{num} is less than the number of colors in the palette,
then only the first \code{num} colors will be displayed.
If \code{num} is greater than the number of colors in the palette,
then that many colors will be generated by linear interpolation
over the vector of colors in the chosen palette.
}
\examples{
viz_palette(mn_palettes$accent)
viz_palette(mn_palettes$accent, "accent: first 4", num = 4)
viz_palette(mn_palettes$accent, "accent: interpolated to 25", num = 25)

}
