% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_mn}
\alias{scale_color_mn}
\alias{scale_colour_mn}
\title{Setup colour palette for ggplot2}
\usage{
scale_color_mn(
  palette = "primary_extended",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)

scale_colour_mn(
  palette = "primary_extended",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'mn_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colours?}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
Setup colour palette for ggplot2
}
\examples{
library(tidyverse)

ggplot(mpg, aes(hwy)) +
  geom_bar(aes(fill = class)) +
  scale_fill_mn(palette = "accent") 
  
ggplot(mtcars, aes(wt, mpg)) +
  geom_point(aes(color = hp), size = 5, alpha = 0.9) +
  scale_color_mn(palette = "blue", discrete = FALSE, reverse = TRUE)
  
ggplot(mpg) +
  geom_point(aes(x = cty, y = hwy, color = class), size = 5, alpha = 0.7) +
  scale_color_mn(palette = "primary_accent")
  
ggplot(diamonds) + 
  geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_mn("primary_extended")
  
}
